Try resources file

+++ Sources/SwaggerSwift/start.swift
@@ -21,22 +21,22 @@ func start(swaggerFilePath: String, token: String, destinationPath: String, proj
 
     let (sourceDirectory, testDirectory) = try! createSwiftProject(at: destinationPath, named: projectName)
 
-    let templateDirectory = Bundle.main.bundleURL.appendingPathComponent("Templates")
-    print("Copying from \(templateDirectory)")
-    for file in try FileManager.default.contentsOfDirectory(at: templateDirectory, includingPropertiesForKeys: nil, options: []) {
-        let cwd = URL(string: FileManager.default.currentDirectoryPath)!
-        let destination: URL
-        if file.absoluteString.contains("Test") {
-            destination = URL(fileURLWithPath: "\(testDirectory)/\(file.lastPathComponent)", relativeTo: cwd).deletingPathExtension().appendingPathExtension("swift")
-        } else {
-            destination = URL(fileURLWithPath: "\(sourceDirectory)/\(file.lastPathComponent)", relativeTo: cwd).deletingPathExtension().appendingPathExtension("swift")
-        }
+    if let templateDirectory = Bundle.main.executableURL?.deletingLastPathComponent().appendingPathComponent("Templates") {
+        for file in try FileManager.default.contentsOfDirectory(at: templateDirectory, includingPropertiesForKeys: nil, options: []) {
+            let cwd = URL(string: FileManager.default.currentDirectoryPath)!
+            let destination: URL
+            if file.absoluteString.contains("Test") {
+                destination = URL(fileURLWithPath: "\(testDirectory)/\(file.lastPathComponent)", relativeTo: cwd).deletingPathExtension().appendingPathExtension("swift")
+            } else {
+                destination = URL(fileURLWithPath: "\(sourceDirectory)/\(file.lastPathComponent)", relativeTo: cwd).deletingPathExtension().appendingPathExtension("swift")
+            }
 
-        if FileManager.default.fileExists(atPath: destination.path) {
-            try FileManager.default.removeItem(at: destination)
-        }
+            if FileManager.default.fileExists(atPath: destination.path) {
+                try FileManager.default.removeItem(at: destination)
+            }
 
-        try FileManager.default.copyItem(at: file, to: destination)
+            try FileManager.default.copyItem(at: file, to: destination)
+        }
     }
 
     if let globalHeaderFields = swaggerFile.globalHeaders?.map({
